Update to general app features
Added a button to the QR Scanner scene in the Menu scene. Also added
input fields to allow IP Address and player number to be easily
configured (for testing). Also created a static class that can be used
to house global variables that need to be accessed between scenes. Also
reverted the Unity version back to 5.5 after accidently updating it to
5.5.1.