Updated controls scene
Changed the controls scene to only show the button that the current
player can use as well as an image to help the user know which direction
they can move the submarine. Also introduced navigation back to the main
menu. Also introduced a new packet type to tell the submarine when the
user has stopped holding the move button so it can stop.